Switch from createTracer to startSegment
startSegment actually works with promises so a lot of the code in
createTracer was able to go away.  Also, instead of forcing you to call
a function to finish the segment it just finishes it when the function
passed in's returned promise resolves.  So all our function needs to do
is return a function that calls startSegment with the original function
passed in.