Update newrelic to version 7 and above
This shouldn't impact any repos using this - only the Herald and CMS use this repo currently, and both are pinned to version 1.1.1 of this repo... which probably means we could upgrade them to a newer version if we wanted, too.

Also upgraded the vagrantfile to use a newer version of Ubuntu, although when I do vagrant up it doesn't work here.  So I just did an npm install from my Mac running node 12.18.3 to get the update.